@@ -3,17 +3,15 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "husky": "^1.3.1",
+    "husky": "^2.2.0",
     "lint-staged": "^8.1.5",
     "prettier": "^1.17.0",
-    "react": "^16.4.2",
+    "react": "^16.8.6",
     "react-dom": "^16.4.2",
     "react-google-maps": "^9.4.5",
-    "react-redux": "^5.0.7",
     "react-router-dom": "^5.0.0",
     "react-scripts": "1.1.4",
-    "recompose": "^0.30.0",
-    "redux": "^4.0.0"
+    "recompose": "^0.30.0"
   },
   "lint-staged": {
     "src/**/*.{js,jsx,json,css}": [
